+++ docs/releases.rst
@@ -5,6 +5,7 @@ Version 1.1, Released ????? 2014
 --------------------------------
 
 * **tappy** has backported support for Python 2.6.
+* **tappy** has support for Python 3.2 and 3.3.
 
 Version 1.0, Released March 2014
 --------------------------------